+++ packages/babel-minify/package.json
@@ -10,7 +10,7 @@
   },
   "dependencies": {
     "babel-core": "^6.9.1",
-    "babel-plugin-conditional-compile": "^0.0.4",
+    "babel-plugin-transform-conditionals": "^0.0.1",
     "babel-plugin-transform-dead-code-elimination": "^2.0.1",
     "babel-plugin-transform-evaluate": "^0.1.3",
     "babel-plugin-transform-mangle": "^0.1.3",
